refactor: refactor pages import
