+++ frontend/src/components/authentication/signup-form.tsx
@@ -34,7 +34,7 @@ export const SignupForm = () => {
         <form onSubmit={handleSubmit(onSubmit)} className="w-full flex flex-col gap-4">
             { isError &&
                 <ErrorMessage
-                message={error.response.data.message ?? 'Something went wrong. Please try again later'}
+                message={error.response?.data?.message ?? 'Something went wrong. Please try again later'}
             />
             }
            <FormField<SignUpSchemaType, 'email'>
