+++ frontend/src/pages/index.ts
@@ -1,3 +1,6 @@
-export * from './signup.page.tsx';
-export * from './login.page.tsx';
+export * from './authentication/signup.page.tsx';
+export * from './authentication/login.page.tsx';
+export * from "./authentication/change-password.page.tsx"
+export * from "./authentication/reset-password-email.page.tsx"
+export * from "./authentication/success-reset-email.page.tsx"
 export * from './not-found.page.tsx'
